+++ base/util.jl
@@ -320,6 +320,10 @@ println_with_color(color::Symbol, msg::AbstractString...) =
 
 ## warnings and messages ##
 
+function bold_if_havecolor(str)
+    have_color ? "\e[1m" * str * "\e[22m" : str
+end
+
 """
     info(msg...; prefix="INFO: ")
 
@@ -336,10 +340,10 @@ julia> info("hello world"; prefix="MY INFO: ")
 MY INFO: hello world
 ```
 """
-function info(io::IO, msg...; prefix="INFO: ")
+function info(io::IO, msg...; prefix= bold_if_havecolor("INFO: "))
     println_with_color(info_color(), io, prefix, chomp(string(msg...)))
 end
-info(msg...; prefix="INFO: ") = info(STDERR, msg..., prefix=prefix)
+info(msg...; prefix=bold_if_havecolor("INFO: ")) = info(STDERR, msg..., prefix=prefix)
 
 # print a warning only once
 
@@ -349,7 +353,7 @@ warn_once(io::IO, msg...) = warn(io, msg..., once=true)
 warn_once(msg...) = warn(STDERR, msg..., once=true)
 
 function warn(io::IO, msg...;
-              prefix="WARNING: ", once=false, key=nothing, bt=nothing,
+              prefix=bold_if_havecolor("WARNING: "), once=false, key=nothing, bt=nothing,
               filename=nothing, lineno::Int=0)
     str = chomp(string(msg...))
     if once
@@ -377,16 +381,16 @@ Display a warning. Argument `msg` is a string describing the warning to be displ
 """
 warn(msg...; kw...) = warn(STDERR, msg...; kw...)
 
-warn(io::IO, err::Exception; prefix="ERROR: ", kw...) =
+warn(io::IO, err::Exception; prefix=bold_if_havecolor("ERROR: "), kw...) =
     warn(io, sprint(buf->showerror(buf, err)), prefix=prefix; kw...)
 
-warn(err::Exception; prefix="ERROR: ", kw...) =
+warn(err::Exception; prefix=bold_if_havecolor("ERROR: "), kw...) =
     warn(STDERR, err, prefix=prefix; kw...)
 
-info(io::IO, err::Exception; prefix="ERROR: ", kw...) =
+info(io::IO, err::Exception; prefix=bold_if_havecolor("ERROR: "), kw...) =
     info(io, sprint(buf->showerror(buf, err)), prefix=prefix; kw...)
 
-info(err::Exception; prefix="ERROR: ", kw...) =
+info(err::Exception; prefix=bold_if_havecolor("ERROR: "), kw...) =
     info(STDERR, err, prefix=prefix; kw...)
 
 function julia_cmd(julia=joinpath(JULIA_HOME, julia_exename()))